requirements.txt
beautifulsoup4==4.11.1
postmarker==1.0
protobuf==6.30.0
python-dotenv==1.0.1
Requests==2.32.3
google-generativeai==0.8.4 
postmarker==1.0